requirements.txt
numpy 
beautifulsoup4 
sentence-transformers 
scikit-learn 
emoji
streamlit
groq
python-dotenv